# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tripio",
        "description": "Документация API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reservings": {
            "get": {
                "tags": [
                    "Reservings"
                ],
                "summary": "Вывод списка записей (index())",
                "description": "Доступ только авторизованным пользователям",
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    },
                    "200": {
                        "description": "Everything OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservingSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservings"
                ],
                "summary": "Добавление новой записи (store())",
                "description": "Возвращает новый созданный route",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservingSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/reservings/{id}": {
            "get": {
                "tags": [
                    "Reservings"
                ],
                "summary": "Вывод конкретной записи (show())",
                "description": "Returns reserving data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Route id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservingSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reservings"
                ],
                "summary": "Обновление записи (update())",
                "description": "Returns updated reserving data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reserving id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReservingSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservingSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reservings"
                ],
                "summary": "Удаление записи (delete())",
                "description": "Deletes a record and returns no content",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reserving id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ReservingSchema": {
                "properties": {
                    "id": {
                        "description": "Reserving ID",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "person_id": {
                        "description": "Foreign key person",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "room_id": {
                        "description": "Foreign key room",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "type_food": {
                        "description": "Тип питания в отеле",
                        "type": "enum",
                        "example": "BB",
                        "nullable": false
                    },
                    "arrival_time": {
                        "description": "День прибытия",
                        "type": "string",
                        "example": "date",
                        "nullable": false
                    },
                    "departure_time": {
                        "description": "День отбытия",
                        "type": "string",
                        "example": "date",
                        "nullable": false
                    },
                    "created_at": {
                        "description": "Date of route creation",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    },
                    "updated_at": {
                        "description": "Date of last updating route data",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}